+++ src/clj/reports/routes/services.clj
@@ -2,6 +2,7 @@
   (:require
    [clojure.java.io :as io]
    [clojure.java.shell :refer [sh]]
+   [clojure.tools.logging :as log]
    [hato.client :as hc]
    [reports.config :refer [env]]
    [reports.db.core :as db]
@@ -19,10 +20,6 @@
 (defn mkdir-p [dir]
   (sh "mkdir" "-p" dir))
 
-(defn copy! [src dest]
-  (println "copy " src " " dest)
-  (io/copy src dest))
-
 ;; destructuring
 (defn upload!
   "受け取った multiplart-params を login/{id}/filename にセーブする。
@@ -30,17 +27,16 @@
   [{{:strs [type login upload]} :multipart-params :as request}]
   (let [{:keys [filename tempfile size]} upload
         dir (dest-dir login type)]
-    (println login type filename tempfile size)
-    (println dir)
-    (mkdir-p dir)
+    (log/debug login type filename tempfile size)
+    (log/debug dir)
     (try
+      (mkdir-p dir)
       (when (empty? filename)
         (throw (Exception. "did not select a file.")))
-      (copy! tempfile (str dir "/" filename))
+      (io/copy tempfile (io/file (str dir "/" filename)))
       (response/ok {:status 200 :body "under construction"})
       (catch Exception e
         (layout/render [request] "error.html" {:message (.getMessage e)})))))
-
 
 (defn services-routes []
   ["/api"